% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestingFunction.R
\name{test_function}
\alias{test_function}
\title{An example for writing a function in R using Roxygen (e.g. a variant of Doxygen for
R) 
Well, this function is for writing stuff, I suppose that I need to write here the
pourpose of the function.}
\usage{
test_function(something = TRUE)
}
\arguments{
\item{something}{Something to put}
}
\description{
An example for writing a function in R using Roxygen (e.g. a variant of Doxygen for
R) 
Well, this function is for writing stuff, I suppose that I need to write here the
pourpose of the function.
}
\examples{
test_function()
}
\keyword{Tests,}
\keyword{examples,}
\keyword{many}
